Move dark_css inside global_css module for css ordering
